html fixes - done;

+++ database/migrations/2021_01_30_155642_create_employees_table.php
@@ -23,7 +23,7 @@ class CreateEmployeesTable extends Migration
             $table->string('phone_number');
             $table->unsignedBigInteger('department_id');
             $table->string('job_position');
-            $table->decimal('salary', 10, 2);
+            $table->integer('salary');
             $table->timestamps();
 
             $table->foreign('department_id')->references('id')->on('departments');
